feat: api generation
