+++ .changeset/open-otters-eat.md
@@ -1,0 +1,5 @@
+---
+'@seedcord/api-docs': minor
+---
+
+api json generator using typedoc
