+++ packages/api-docs/eslint.config.mjs
@@ -1,0 +1,12 @@
+import createConfig from '@seedcord/eslint-config';
+
+export default createConfig({
+  tsconfigRootDir: import.meta.dirname,
+  userConfigs: [
+    {
+      rules: {
+        'no-console': 'off'
+      }
+    }
+  ]
+});
